fix(packets): send a changed game point in the shape the client parses
The delta carried a leading entry count the client never reads, so it parsed no
entries at all: a point change never reached a window that was already open, and
the character sheet kept its old numbers until the window was reopened and re-read
the cache that SCCharacterGamePointsPacket fills.

Verified against the 10.0.2.13 client by sending raw payloads. A bare
{u16 kind, u32 amount} pair with kind 1 and amount 2002 prints "Acquired 2,002
Honor Points" and moves the cached value by exactly 2002. Kinds 0 and 2..52 are
dropped without a message, a point change or an event, so honor is the only point
this opcode can announce.

Every other point therefore rides SCCharacterGamePointsPacket plus a zero-amount
honor delta, whose only job is to raise the event an open window repaints on. The
two packets are ordered by kind, because the client behaves in two conflicting
ways: it prints "cached + amount" for the delta, which needs the cache to still
hold the pre-change number, while a window that shows game points re-reads every
currency when the event fires, which needs the table to have landed first. Honor
sends the delta first, everything else the table first.

Also adds the vocation point command the change was tested with.

+++ AAEmu.Game/Core/Packets/G2C/SCGamePointChangedPacket.cs
@@ -3,15 +3,23 @@ using AAEmu.Game.Core.Network.Game;
 
 namespace AAEmu.Game.Core.Packets.G2C;
 
+/// <summary>
+/// Delta notification for the character sheet's game-point table. Display-only -
+/// SCCharacterGamePointsPacket carries the values, this is what makes an already-open sheet repaint.
+///
+/// Wire format is a bare {u16 kind, u32 amount} - there is no entry count. Verified against the
+/// 10.0.2.13 client: it reads the kind as a u16 and the amount as a u32, then prints "cached + amount",
+/// so the delta has to be applied while the client's cache still holds the pre-change number.
+///
+/// Kind space: this client accepts only kind 1 (honor) on this opcode. Kinds 0 and 2..52 are silently
+/// dropped - no chat line, no point change, no event - so every other point is announced by
+/// SCCharacterGamePointsPacket plus a zero-amount honor delta whose only job is to raise that event.
+/// </summary>
 public class SCGamePointChangedPacket(byte kind, int amount) : GamePacket(SCOffsets.SCGamePointChangedPacket, 1)
 {
-    // TODO kind:
-    // 0 - honor
-    // 1 - vocation(living)
-
     public override PacketStream Write(PacketStream stream)
     {
-        stream.Write(kind);
+        stream.Write((ushort)kind);
         stream.Write(amount);
         return stream;
     }
